i8 for my_num_vars
fixes #1051

+++ assimilation_code/location/channel/location_mod.f90
@@ -1353,7 +1353,7 @@ subroutine convert_vertical_obs(ens_handle, num, locs, loc_kinds, loc_types, &
                                 which_vert, status)
 
 type(ensemble_type), intent(in)    :: ens_handle
-integer,             intent(in)    :: num
+integer(i8),         intent(in)    :: num
 type(location_type), intent(inout) :: locs(:)
 integer,             intent(in)    :: loc_kinds(:), loc_types(:)
 integer,             intent(in)    :: which_vert
@@ -1369,7 +1369,7 @@ subroutine convert_vertical_state(ens_handle, num, locs, loc_kinds, loc_indx, &
                                   which_vert, istatus)
 
 type(ensemble_type), intent(in)    :: ens_handle
-integer,             intent(in)    :: num
+integer(i8),         intent(in)    :: num
 type(location_type), intent(inout) :: locs(:)
 integer,             intent(in)    :: loc_kinds(:)
 integer(i8),         intent(in)    :: loc_indx(:)
